+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  -   Add "Stop All CF applications for the specified org name" action
  -   Add "Start a CF application" action
+ -   Add "Start All CF applications for the specified org name" action
 
 ## [0.6.0][]